[Feat/#22] setting message test env (room.html)
