+++ src/main/java/com/messenger/chatty/domain/message/controller/MessageController.java
@@ -44,6 +44,7 @@ public class MessageController {
         template.convertAndSend("amq.topic", "channel." + channelIdStr, messageDto);
     }
 
+
     // receiver()는 단순히 큐에 들어온 메세지를 소비만 한다. (현재는 디버그 용도)
     @RabbitListener(queues = "message.queue")
     public void receive(MessageDto messageDto) {
